+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add in Get Help button to more instances when available [#734]
 - Add dialog warning user when trying to skip external download mods from CurseForge
 - Add logging for Java version and install method on launch if analytics enabled
+- Log warnings when useSystemGlfw/useSystemOpenAl are enabled when launching
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg